# apps/api/openapi.json
{"openapi":"3.0.0","paths":{"/todo":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoDto"}}}},"responses":{"201":{"description":""}},"tags":["Todo"]},"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Todo"]}},"/todo/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}}},"tags":["Todo"]},"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Todo"]}},"/categories":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":""}},"tags":["Categories"]},"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Categories"]}},"/categories/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Categories"]},"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Categories"]}}},"info":{"title":"Todo API","description":"A todo API with categories","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateTodoDto":{"type":"object","properties":{}},"Category":{"type":"object","properties":{}},"Todo":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"completed":{"type":"boolean"},"createdAt":{"format":"date-time","type":"string"},"completedAt":{"format":"date-time","type":"string","nullable":true},"categories":{"type":"array","items":{"$ref":"#/components/schemas/Category"}}},"required":["id","title","description","completed","createdAt","categories"]},"CreateCategoryDto":{"type":"object","properties":{}}}}}