# apps/web/api-client/src/models/Todo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Todo API
 * A todo API with categories
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Todo
 */
export interface Todo {
    /**
     * 
     * @type {string}
     * @memberof Todo
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Todo
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof Todo
     */
    description: string;
    /**
     * 
     * @type {boolean}
     * @memberof Todo
     */
    completed: boolean;
    /**
     * 
     * @type {Date}
     * @memberof Todo
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Todo
     */
    completedAt?: Date | null;
    /**
     * 
     * @type {Array<object>}
     * @memberof Todo
     */
    categories: Array<object>;
}

/**
 * Check if a given object implements the Todo interface.
 */
export function instanceOfTodo(value: object): value is Todo {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('completed' in value) || value['completed'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('categories' in value) || value['categories'] === undefined) return false;
    return true;
}

export function TodoFromJSON(json: any): Todo {
    return TodoFromJSONTyped(json, false);
}

export function TodoFromJSONTyped(json: any, ignoreDiscriminator: boolean): Todo {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'title': json['title'],
        'description': json['description'],
        'completed': json['completed'],
        'createdAt': (new Date(json['createdAt'])),
        'completedAt': json['completedAt'] == null ? undefined : (new Date(json['completedAt'])),
        'categories': json['categories'],
    };
}

export function TodoToJSON(json: any): Todo {
    return TodoToJSONTyped(json, false);
}

export function TodoToJSONTyped(value?: Todo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'description': value['description'],
        'completed': value['completed'],
        'createdAt': ((value['createdAt']).toISOString()),
        'completedAt': value['completedAt'] == null ? undefined : ((value['completedAt'] as any).toISOString()),
        'categories': value['categories'],
    };
}